ported MineTweaker integration to new compat system
