--- src/main/java/buildcraftAdditions/ModIntegration/MineTweaker/MineTweakerIntegreation.java
@@ -1,19 +1,0 @@
-package buildcraftAdditions.ModIntegration.MineTweaker;
-
-import minetweaker.MineTweakerAPI;
-
-/**
- * Copyright (c) 2014, AEnterprise
- * http://buildcraftadditions.wordpress.com/
- * Buildcraft Additions is distributed under the terms of GNU GPL v3.0
- * Please check the contents of the license located in
- * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
- */
-public class MineTweakerIntegreation {
-
-	public static void integrate() {
-		MineTweakerAPI.registerClass(Dusters.class);
-		MineTweakerAPI.registerClass(Refinery.class);
-		MineTweakerAPI.registerClass(Cooling.class);
-	}
-}
