+++ src/main/java/buildcraftAdditions/compat/minetweaker/script/Cooling.java
@@ -1,4 +1,4 @@
-package buildcraftAdditions.ModIntegration.MineTweaker;
+package buildcraftAdditions.compat.minetweaker.script;
 
 import buildcraftAdditions.api.recipe.BCARecipeManager;
 
@@ -10,34 +10,41 @@ import minetweaker.api.minecraft.MineTweakerMC;
 import stanhebben.zenscript.annotations.ZenClass;
 import stanhebben.zenscript.annotations.ZenMethod;
 
-@ZenClass("mods.bcadditions.Refinery")
+/**
+ * Copyright (c) 2014, AEnterprise
+ * http://buildcraftadditions.wordpress.com/
+ * Buildcraft Additions is distributed under the terms of GNU GPL v3.0
+ * Please check the contents of the license located in
+ * http://buildcraftadditions.wordpress.com/wiki/licensing-stuff/
+ */
+@ZenClass("mods.bcadditions.Cooling")
 @ModOnly("bcadditions")
-public class Refinery {
+public class Cooling {
 
 	@ZenMethod
-	public static void addRefineryRecipe(ILiquidStack input, ILiquidStack output, int requiredHeat) {
-		MineTweakerAPI.apply(new AddRecipeAction(input, output, requiredHeat));
+	public static void addCoolingRecipe(ILiquidStack input, ILiquidStack output, float heat) {
+		MineTweakerAPI.apply(new AddRecipeAction(input, output, heat));
 	}
 
 	@ZenMethod
-	public static void removeRefineryRecipe(ILiquidStack input) {
+	public static void removeCoolingRecipe(ILiquidStack input) {
 		MineTweakerAPI.apply(new RemoveRecipeAction(input));
 	}
 
 	private static class AddRecipeAction implements IUndoableAction {
 		private final ILiquidStack input;
 		private final ILiquidStack output;
-		private final int requiredHeat;
+		private final float heat;
 
-		public AddRecipeAction(ILiquidStack input, ILiquidStack output, int requiredHeat) {
+		public AddRecipeAction(ILiquidStack input, ILiquidStack output, float heat) {
 			this.input = input;
 			this.output = output;
-			this.requiredHeat = requiredHeat;
+			this.heat = heat;
 		}
 
 		@Override
 		public void apply() {
-			BCARecipeManager.refinery.addRecipe(MineTweakerMC.getLiquidStack(input), MineTweakerMC.getLiquidStack(output), requiredHeat);
+			BCARecipeManager.cooling.addRecipe(MineTweakerMC.getLiquidStack(input), MineTweakerMC.getLiquidStack(output), heat);
 		}
 
 		@Override
@@ -47,17 +54,17 @@ public class Refinery {
 
 		@Override
 		public void undo() {
-			BCARecipeManager.refinery.removeRecipe(MineTweakerMC.getLiquidStack(input));
+			BCARecipeManager.cooling.removeRecipe(MineTweakerMC.getLiquidStack(input));
 		}
 
 		@Override
 		public String describe() {
-			return String.format("Adding BCA Refinery recipe for %s -> %s", input, output);
+			return String.format("Adding BCA Cooling Tower recipe for %s -> %s", input, output);
 		}
 
 		@Override
 		public String describeUndo() {
-			return String.format("Removing BCA Refinery recipe for %S -> %s", input, output);
+			return String.format("Removing BCA Cooling Tower recipe for %s -> %s", input, output);
 		}
 
 		@Override
@@ -75,7 +82,7 @@ public class Refinery {
 
 		@Override
 		public void apply() {
-			BCARecipeManager.refinery.removeRecipe(MineTweakerMC.getLiquidStack(input));
+			BCARecipeManager.cooling.removeRecipe(MineTweakerMC.getLiquidStack(input));
 		}
 
 		@Override
@@ -90,7 +97,7 @@ public class Refinery {
 
 		@Override
 		public String describe() {
-			return String.format("Removing BCA Refinery recipe for %S -> %s", input, BCARecipeManager.refinery.getRecipe(MineTweakerMC.getLiquidStack(input)).getOutput().getLocalizedName());
+			return String.format("Removing BCA Cooling Tower recipe for %s -> %s", input, BCARecipeManager.cooling.getRecipe(MineTweakerMC.getLiquidStack(input)).getOutput().getLocalizedName());
 		}
 
 		@Override
